refactor(storage): split CIMD tests by concern
